rename startControls
